Summary of Commits for Starting a New Game
All test cases in test suite pass. I deleted source code from my
components that I hadn't tested first and started over so that this time
I implemented TDD first before writing my source code. Modified the
source code accordingly as I worked through the green/red TDD cycle. The
empty board is now created and renders to a new page at the /board url.
Heading for the players info renders at the /game url. Currently working
on a way to render both of these components to the /game url. Also
modified the header component so that the text for player 1 and player 2
is evenly distributed across 2 columns.

Second story is completed. Another page is shown when the start game is
clicked. On the second page the user is able to see an empty tic
tac toe board in which the spaces on the board are simply div tags
styled with css. The heading for the players info is also shown on the same
page. All tests in the test suite pass. There are currently some tests
which test static components. I will remove these if it's decided that
they should removed. All tests in the test suite pass. Also created a game
component that encompasses all of the components for the app.

Today I also modified the test suite to include two more tests which
will find  the css class selector of board-grid for the board component
and the css id selector of space for the square component. All tests in test suite  pass.

Updated working tree app.test.js

Removed debugging logs to console

Edited the game page so that it only renders one component for the
entire page.

I decided to remove the game component so that I do not have
duplicate code and let the game.js file encompass all the code needed
to render to the second page via react router.

Modified test suite and removed test case on game component

Applied all changes recommended from the pull request first review (by
Andy). I separated my tests into a more organized and clean file structure so the
tests are easier to read.

Modified code so that it now includes the recommended changes from the latest PR